Mission Editing

+++ api/web/src/components/CloudTAK/util/Feature.vue
@@ -22,13 +22,34 @@
         }'
         @click='flyTo'
     >
-        <IconMapPin
-            :size='20'
-            :stroke='1'
-            class='me-2'
-        />
+        <span class='me-2'>
+            <IconLine
+                v-if='feature.geometry && feature.geometry.type === "LineString"'
+                :size='20'
+                :color='feature.properties.stroke || "white"'
+                stroke='1'
+            />
+            <IconCone
+                v-else-if='feature.properties && feature.properties.sensor'
+                :size='20'
+                :color='feature.properties.stroke || "white"'
+                stroke='1'
+            />
+            <IconPolygon
+                v-else-if='feature.geometry && feature.geometry.type === "Polygon"'
+                :size='20'
+                :color='feature.properties.fill || "white"'
+                stroke='1'
+            />
+            <IconMapPin
+                v-else
+                :size='20'
+                stroke='1'
+            />
+        </span>
         <div
             class='text-truncate user-select-none'
+            style='width: 180px;'
             v-text='feature.properties.callsign || feature.properties.name || "Unnamed"'
         />
 
@@ -37,8 +58,16 @@
             class='ms-auto btn-list hover-button-hidden'
         >
             <TablerDelete
+                v-if='deleteAction === "delete"'
                 :size='20'
                 displaytype='icon'
+                @delete='deleteCOT'
+            />
+            <IconTrash
+                v-else
+                :size='20'
+                class='cursor-pointer'
+                :stroke='1'
                 @click='deleteCOT'
             />
         </div>
@@ -51,7 +80,11 @@ import {
     TablerDelete
 } from '@tak-ps/vue-tabler';
 import {
-    IconMapPin
+    IconMapPin,
+    IconTrash,
+    IconLine,
+    IconCone,
+    IconPolygon,
 } from '@tabler/icons-vue';
 import { useCOTStore } from '/src/stores/cots.ts';
 const cotStore = useCOTStore();
@@ -63,7 +96,11 @@ export default {
     components: {
         Contact,
         TablerDelete,
-        IconMapPin
+        IconMapPin,
+        IconTrash,
+        IconLine,
+        IconCone,
+        IconPolygon,
     },
     props: {
         feature: {
@@ -77,6 +114,10 @@ export default {
             type: Boolean,
             default: true
         },
+        deleteAction: {
+            type: String,
+            default: 'delete' //emit or delete
+        },
         hover: {
             type: Boolean,
             default: true
@@ -86,6 +127,7 @@ export default {
             default: true
         }
     },
+    emits: ['delete'],
     computed: {
         isZoomable: function() {
             if (this.mission) {
@@ -99,7 +141,11 @@ export default {
     },
     methods: {
         deleteCOT: async function() {
-            await cotStore.delete(this.feature.id);
+            if (this.deleteAction === 'delete') {
+                await cotStore.delete(this.feature.id);
+            } else {
+                this.$emit('delete');
+            }
         },
         flyTo: function() {
             if (!this.isZoomable) return;
